Initial commit based on Bolt.New AI Prompt
Create a web application that will be used to display dashboard information for Sitecore XM Cloud Users. The application should use the styles and components that are available on https://blok.sitecore.com/
This should use React version 18, Chakra UI v2, and Material Design Library.
The dashboard should include the following information:
Most recent publish job (job, person who performed publish, publishing options, date and time of publish, publish stats which includes number of items published, skipped, and failed)
Total number of website collections and within each collection the number of websites that belong to the collection. For each website include the number of Pages on the website.
Total number of assets that are in their Media Library